Side mav styling and cleanup
apparently the icon issue was fixed by me adding that line to index.html.

i added an icon for the side nav menu. also, i added a close icon for it. let me know what you think about it. We can change it or go back to text.

I was unable to get that side nav menu to look somewhat decent. i couldn't seem to override it with a new css class. let me know the fix and ill work on it. thanks!